SoftcamManager: Limit VM usage
uses lots of threads and uses the default 8MB stack size for each.
This makes it allocate >100MB address space. By limiting the stack size
to 512k the VM usage is recuced, and if it uses so much stack, something's
wrong with it anyway so better kill the CAM instead of causing panic.